pyproject.toml
[project]
name = "stockmarket"
version = "0.0.1"
description = "Stock market Analysis Toolkit"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.128.0",
    "langchain>=1.2.0",
    "langchain-google-genai>=4.1.2",
    "playwright>=1.57.0",
    "playwright-stealth>=2.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "setuptools>=80.9.0",
    "uvicorn[standard]>=0.40.0",
]
